Further simplify the package
The core module required the `__version__` attribute, which really
belongs in the `__init__.py` module file, but the latter imports from
core.py so that would introuce a circular import. The attribute could be
defined in the core.py file and imported into __init__.py, but this
reduces our ability to fetch the version attribute through means other
than Python imports.

Since this is effectively a one file module anyhow, flattening the
module is a good solution.